mypage, otherpage profile tabs 합치고 경로 변경
