+++ src/components/user/Tabs/UploadProject.js
@@ -1,23 +1,20 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { Cookies } from "react-cookie";
-import { useParams } from "react-router-dom";
-import ProjectCards from "../../../ui/project/ProjectCards";
-import classes from "../mytabs.module.css"
+import ProjectCards from "../../ui/project/ProjectCards";
 
-const SupportProject = () => {
-    const [projects, setProjects] = useState([]);
+import classes from "../mypage.module.css"
 
-    let { params } = useParams()
+const UploadProject = () => {
+    const [projects, setProjects] = useState([])
+    const cookies = new Cookies()
+    const token = cookies.get('user_token')
 
-    const cookie = new Cookies()
-    const token = cookie.get('user_token')
-
-    const findBuy = () => {
-        axios.get(`http://localhost:3000/u/${params}/buy`, token ? { headers : {'user_token' : token} } : '')
-        .then(res => {
-            console.log(res.data)
-            setProjects(res.data)
+    const findPost = () => {
+        axios.get("http://localhost:3000/mypage/upload" ,{headers : {'user_token': token}})
+        .then(response => {
+            console.log(response.data)
+            setProjects(response.data)
         })
         .catch(e => {
             console.log(e)
@@ -25,16 +22,14 @@ const SupportProject = () => {
     }
 
     useEffect(() => {
-        findBuy()
+        findPost();
     }, [])
-    
-    return (
-        <>
+
+    return(
+        <div>
             {projects.length !== 0 ?
                 <div>
-                    <div className={classes.upprolength}>
-                        <span>{projects.length}</span>개의 프로젝트가 있습니다.
-                    </div>
+                    <div className={classes.upprolength}><span>{projects.length}</span>개의 프로젝트가 있습니다.</div>
                     <div className={classes.upprojectbox}>
                         {projects.map((prol, key) => (
                             <div key={key}>
@@ -43,14 +38,14 @@ const SupportProject = () => {
                         ))}
                     </div>
                 </div>
-            :
+                :
                 <div>
                     <div className={classes.upprolength}><span>{projects.length}</span>개의 프로젝트가 있습니다.</div>
                     <div>올린 프로젝트가 없습니다</div>
                 </div>
             }
-        </>
+        </div>
     )
 }
 
-export default SupportProject
+export default UploadProject
