--- src/components/user/Tabs/mytabs.module.css
@@ -1,78 +1,0 @@
-.tabsbox {
-    padding: 0px !important;
-}
-
-.tabsbox .tabscontent {
-    max-width: 1160px;
-    margin: 0 auto;
-}
-
-.tabsbox .tabitem {
-    max-width: 1184px;
-    margin: 0 auto;
-}
-
-/* .tabsbox .tabitem .tabitemTitle:first-child {
-    margin: 0 -12px;
-} */
-
-.tabsbox .tabitem .tabitemTitle {
-    font-size: 18px;
-}
-
-/* ProfileTab */ 
-
-.intro {
-    padding: 20px 0px;
-}
-
-/* uploadProject */
-
-.upprolength {
-    padding-left: 12px;
-    padding: 14px 12px 24px;
-    margin: 0px -12px
-}
-.upprolength span {
-    color: #ff5757;
-}
-
-.upprojectbox {
-    margin: 0px -12px;
-    display: flex;
-    flex-wrap: wrap;
-}
-
-/* profile */
-
-.profilebox {
-    max-width: 1160px;
-    margin: 0 auto;
-    display: flex;
-    align-items: center;
-    padding: 40px 0 24px;
-}
-
-.profilebox img {
-    width: 80px;
-    height: 80px;
-    object-fit: cover;
-    border-radius: 50%;
-}
-
-.profilebox .profileInfo {
-    margin-left: 24px;
-}
-
-.profilebox .profileInfo div:first-child {
-    color: #1c1c1c;
-    font-weight: bold;
-    font-size: 20px;
-}
-
-.profilebox .profileInfo div:nth-child(2) {
-    color: #6d6d6d;
-    font-size: 13px;
-}
-
-
